Add files via upload
For Predicate Logic or Sentence Logic
infixtopolish(wff) is a function that takes infix notation wff and outputs Polish notation
polishtoinfix (wff) is a function that takes a Polish notation wff and outputs infix notation
infix_wff_check(wff) check to see if an infix notation sentence is well-formed
wffcheck(wff) checks to see if a Polish notation sentence is well-formed 

operators are : ~, &, v, >, <>, (3x) [existential], (4x) [universal quantifier]. You can also input the unicode characters for the quantifiers. Note that you cannot use "v" as a variable or name because it's used for the disjunction.

Started to add the ability to use astericks after variables or objects to increase number of variable and object names, but haven't finished yet.